Add styling of choose board size page
This commit contains the styling of the choose board size page, but
these new rules also demanded som change on existing rules.
I had to add ids on the sections to be able to keep the duplicated
buttons apart (#start-page and #how-to-play-page).

The so called home page is from now on referred to as start page.