Solve responsiveness issue for start page image
The responsiveness for the start page image on tall screens needs to
be solved by a function, therefore it is done in style.js.

This means the start page needs to be reloaded when the page is
is resized and therefore some changes had to be made on different places
in the js-file to make the eventlistener, already set for the play game
page, be valid also for the start page.